Add binding. Add InitActivity. Rename InitScreen to WelcomeActivity; InitPresenter to WelcomePresenter. Change launch activity
