+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
 
     <application
         android:allowBackup="true"
@@ -11,21 +11,23 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.TravelHelper"
-        tools:targetApi="31">
+        tools:targetApi="31" >
         <activity
-            android:name=".screens.InitScreen"
-            android:exported="false">
-        </activity>
-        <activity
-            android:name=".MainActivity"
-            android:name=".screens.MainActivity"
+            android:name=".screens.InitActivity"
             android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
+        <intent-filter>
+            <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
+            <category android:name="android.intent.category.LAUNCHER" />
+        </intent-filter>
         </activity>
+
+        <activity
+            android:name=".screens.WelcomeActivity"
+            android:exported="false"/>
+        <activity
+            android:name=".screens.MainActivity"
+            android:exported="false" />
         <activity
             android:name=".screens.HotelActivity"
             android:exported="false" />
