+++ app/src/main/java/ru/itmo/travelhelper/presenter/WelcomePresenter.kt
@@ -2,7 +2,7 @@ package ru.itmo.travelhelper.presenter
 
 import ru.itmo.travelhelper.view.InitView
 
-class InitPresenter (private val view: InitView) {
+class WelcomePresenter (private val view: InitView) {
     fun setNextScreen(currentScreenInitNumber: Int) {
         view.showNextImageInit(currentScreenInitNumber)
         view.showNextTitleAndTextInit(currentScreenInitNumber)
