+++ app/src/main/java/ru/itmo/travelhelper/screens/InitActivity.kt
@@ -1,0 +1,26 @@
+package ru.itmo.travelhelper.screens
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+
+class InitActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+
+        val isFirstLaunching = true //get bool value if first launching true/false
+
+        if (isFirstLaunching) {
+            val intentWelcomeActivity = Intent(this, WelcomeActivity::class.java)
+            startActivity(intentWelcomeActivity)
+        }
+        else {
+            val intentMainActivity = Intent(this, MainActivity::class.java)
+            startActivity(intentMainActivity)
+        }
+
+
+    }
+}
