+++ app/src/main/java/ru/itmo/travelhelper/screens/WelcomeActivity.kt
@@ -1,0 +1,64 @@
+package ru.itmo.travelhelper.screens
+
+
+import android.os.Bundle
+import android.widget.RadioButton
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.ContextCompat
+import ru.itmo.travelhelper.R
+import ru.itmo.travelhelper.databinding.ActivityWelcomeScreenBinding
+import ru.itmo.travelhelper.presenter.WelcomePresenter
+import ru.itmo.travelhelper.view.InitView
+
+
+
+class WelcomeActivity : AppCompatActivity(), InitView {
+    private lateinit var welcomePresenter: WelcomePresenter
+
+    private val binding by lazy {
+        ActivityWelcomeScreenBinding.inflate(layoutInflater)
+    }
+
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+
+        welcomePresenter = WelcomePresenter(this)
+
+
+        var currentScreenInitNumber = 0
+        welcomePresenter.setNextScreen(currentScreenInitNumber)
+
+        binding.nextButtonInit.setOnClickListener {
+            if (currentScreenInitNumber < 3) {welcomePresenter.setNextScreen(++currentScreenInitNumber)}
+
+        }
+
+    }
+
+    override fun showNextTitleAndTextInit(initScreenNumber: Int) {
+        val titleArray = resources.getStringArray(R.array.init_array_titles)
+        binding.titleTextInit.text = titleArray[initScreenNumber]
+
+        val textArray = resources.getStringArray(R.array.init_array_texts)
+        binding.mainTextInit.text = textArray[initScreenNumber]
+    }
+
+    override fun showNextImageInit(initScreenNumber: Int) {
+        val imageResource: Int = resources.getIdentifier("@drawable/init_picture_${initScreenNumber+1}", null, packageName)
+        val imageView = ContextCompat.getDrawable(this, imageResource)
+        binding.imageInit.setImageDrawable(imageView)
+    }
+
+
+    override fun showNextButtonTextInit(initScreenNumber: Int) {
+        val buttonArray = resources.getStringArray(R.array.init_array_buttons)
+        binding.nextButtonInit.text = buttonArray[if (initScreenNumber < 3) 0 else 1]
+
+        val radioButtonResource = resources.getIdentifier("radioButton${initScreenNumber+1}_Init", "id", packageName)
+        val radioButton = findViewById<RadioButton>(radioButtonResource)
+        radioButton.isChecked = true
+    }
+
+}
