+++ app/src/main/res/layout/activity_welcome_screen.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/init_main_bg"
-    tools:context=".screens.InitScreen">
+    tools:context=".screens.WelcomeActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/headerConstraintLayout"
