+++ app/src/main/res/values/strings.xml
@@ -34,5 +34,4 @@
     <string name="init_button_start">Начать</string>
 
 
-
 </resources>
